Extracting components in coursepage
